link candidates are accessed via transit route stop instead of stop facility (which can be visited twice in a route)
